Add flexibility to Ghost Hunting
This commit adds flexibility to the Ghost Hunting technique where previously it was rigidly checking for enum values for the origin of where the hooked API came from. Instead, we define what can cancel Ghost Hunt timers and push each event to a vector, which then get removed individually when the corresponding notifications are received.