Receive WriteVirtualMemory signal from hooked DLL
This commit receives and adds WriteVirtualMemory events from the hooked syscall in malware.exe.

A bug has been found in the driver in that two callback routines seem to run for a process handle being intercepted for a remote process - this is odd as it didn't occur in the past. To investigate. This commit removes the monitoring in the EDR of OpenProcess until the bug is fixed.